+++ src/components/cdns/Index.vue
@@ -29,6 +29,7 @@
                 </q-item>
               </VirtualList>
             </q-list>
+            <q-btn icon="mdi-download" class="deleted-action" @click="getDeletedHandler" v-if="needShowGetDeletedAction && tokenType === 1">get deleted</q-btn>
           </q-popover>
         </q-btn>
         <div v-if="!items.length">{{isLoading ? 'Fetching data..' : 'CDN`s not found'}}</div>
@@ -38,7 +39,7 @@
       <q-toolbar slot="header" color="dark" class="justify-between">
         <q-item class="no-padding" style="max-width: 50%" :style="{cursor: isNeedSelect ? '' : 'default!important'}">
           <q-item-main>
-            <q-item-tile label class="ellipsis overflow-hidden" :style="{maxWidth: '140px'}">{{selectedItem.name}}</q-item-tile>
+            <q-item-tile label class="ellipsis overflow-hidden" :style="{maxWidth: '140px'}">{{selectedItem.name || '&lt;noname&gt;'}}</q-item-tile>
           </q-item-main>
           <q-item-side class="text-right">
             <q-item-tile style="display: inline-block" stamp color="white" class="text-center"><div v-if="selectedItem.deleted" class="cheap-modifier"><small>DELETED</small></div>#{{selectedItem.id.toString()}}</q-item-tile>
@@ -68,6 +69,7 @@
                 </q-item>
               </VirtualList>
             </q-list>
+            <q-btn icon="mdi-download" class="deleted-action" @click="getDeletedHandler" v-if="needShowGetDeletedAction && tokenType === 1">get deleted</q-btn>
           </q-popover>
         </q-item>
         <q-btn v-if="!selectedItem.deleted" flat class="on-left" color="white" @click="modeModel = !modeModel" :icon="modeModel ? 'playlist_play' : 'history'"  :rounded="$q.platform.is.mobile">
@@ -85,6 +87,7 @@
         ref="logs"
         :mode="mode"
         :item="selectedItem"
+        :limit="limit"
         originPattern="storage/cdns/:id"
         :isEnabled="true"
         :config="config.logs"
@@ -98,7 +101,7 @@
 
 <script>
 import logs from '../logs/Index.vue'
-import { mapState } from 'vuex'
+import { mapState, mapActions } from 'vuex'
 import VirtualList from 'vue-virtual-scroll-list'
 
 export default {
@@ -113,7 +116,8 @@ export default {
     return {
       mode: 1,
       active: null,
-      isInit: false
+      isInit: false,
+      needShowGetDeletedAction: true
     }
   },
   computed: {
@@ -123,7 +127,8 @@ export default {
       },
       isEmptyMessages (state) {
         return state[this.config.logs.vuexModuleName] ? !state[this.config.logs.vuexModuleName].messages.length : false
-      }
+      },
+      tokenType (state) { return state.tokenInfo.access ? state.tokenInfo.access.type : -1 }
     }),
     items () {
       return this.$store.state.items
@@ -144,6 +149,7 @@ export default {
     }
   },
   methods: {
+    ...mapActions(['getDeleted']),
     viewDataHandler (content) {
       this.$emit('view-data', content)
     },
@@ -158,6 +164,10 @@ export default {
         cancel: true
       }).then(() => { this.$store.commit(`${this.config.logs.vuexModuleName}/clearMessages`) })
         .catch(() => {})
+    },
+    async getDeletedHandler () {
+      await this.getDeleted('cdns')
+      this.needShowGetDeletedAction = false
     }
   },
   created () {
@@ -224,5 +234,13 @@ export default {
     color: white;
     padding: 0 3px;
     margin-bottom: 3px;
+  }
+  .deleted-action {
+    width: 100%;
+    color: #999;
+    background-color: #eee;
+    font-size: .7rem;
+    padding-top: 0;
+    padding-bottom: 0
   }
 </style>