+++ src/components/channels/Index.vue
@@ -31,6 +31,7 @@
                   </q-item>
                 </VirtualList>
               </q-list>
+              <q-btn icon="mdi-download" class="deleted-action" @click="getDeletedHandler" v-if="needShowGetDeletedAction && tokenType === 1">get deleted</q-btn>
             </q-popover>
           </q-btn>
           <div v-if="!items.length">{{isLoading ? 'Fetching data..' : 'Channels not found'}}</div>
@@ -42,7 +43,7 @@
         <q-item class="no-padding" :style="{cursor: isNeedSelect ? '' : 'default!important'}">
           <q-item-main>
             <q-tooltip><small>protocol: {{selectedItem.protocol_name || selectedItem.protocol_id}}</small></q-tooltip>
-            <q-item-tile label class="ellipsis overflow-hidden" :style="{maxWidth: '140px'}">{{selectedItem.name}}</q-item-tile>
+            <q-item-tile label class="ellipsis overflow-hidden" :style="{maxWidth: '140px'}">{{selectedItem.name || '&lt;noname&gt;'}}</q-item-tile>
             <q-item-tile sublabel style="font-size: 0.8rem">{{selectedItem.uri}}</q-item-tile>
           </q-item-main>
           <q-item-side class="text-right">
@@ -75,6 +76,7 @@
                 </q-item>
               </VirtualList>
             </q-list>
+            <q-btn icon="mdi-download" class="deleted-action" @click="getDeletedHandler" v-if="needShowGetDeletedAction && tokenType === 1">get deleted</q-btn>
           </q-popover>
         </q-item>
         <q-btn v-if="!selectedItem.deleted" flat class="on-left" color="white" @click="modeModel = !modeModel" :icon="modeModel ? 'playlist_play' : 'history'" :rounded="$q.platform.is.mobile">
@@ -119,6 +121,7 @@
         ref="logs"
         :mode="mode"
         :item="selectedItem"
+        :limit="limit"
         :isEnabled="!!+size[0]"
         originPattern="gw/channels/:id"
         :config="config.logs"
@@ -145,7 +148,7 @@
 <script>
 import logs from '../logs/Index.vue'
 import messages from './messages/Index.vue'
-import { mapState } from 'vuex'
+import { mapState, mapActions } from 'vuex'
 import VirtualList from 'vue-virtual-scroll-list'
 
 export default {
@@ -161,7 +164,8 @@ export default {
       mode: 1,
       active: null,
       ratio: 50,
-      isInit: false
+      isInit: false,
+      needShowGetDeletedAction: true
     }
   },
   computed: {
@@ -175,7 +179,8 @@ export default {
         let hasntMessages = this.config.messages && state[this.config.messages.vuexModuleName] && !state[this.config.messages.vuexModuleName].messages.length && this.ratio !== 100,
           hasntLogs = this.config.logs && state[this.config.logs.vuexModuleName] && state[this.config.logs.vuexModuleName].messages && !state[this.config.logs.vuexModuleName].messages.length && this.ratio !== 0
         return hasntMessages && hasntLogs
-      }
+      },
+      tokenType (state) { return state.tokenInfo.access ? state.tokenInfo.access.type : -1 }
     }),
     size () {
       return [this.ratio, 100 - this.ratio]
@@ -199,6 +204,7 @@ export default {
     }
   },
   methods: {
+    ...mapActions(['getDeleted']),
     viewDataHandler (content) {
       this.$emit('view-data', content)
     },
@@ -219,6 +225,10 @@ export default {
         this.$store.commit(`${this.config.logs.vuexModuleName}/clearMessages`)
       })
         .catch(() => {})
+    },
+    async getDeletedHandler () {
+      await this.getDeleted('channels')
+      this.needShowGetDeletedAction = false
     }
   },
   created () {
@@ -300,5 +310,13 @@ export default {
     color: white;
     padding: 0 3px;
     margin-bottom: 3px;
+  }
+  .deleted-action {
+    width: 100%;
+    color: #999;
+    background-color: #eee;
+    font-size: .7rem;
+    padding-top: 0;
+    padding-bottom: 0
   }
 </style>