+++ src/components/devices/Index.vue
@@ -29,6 +29,7 @@
                 </q-item>
               </VirtualList>
             </q-list>
+            <q-btn icon="mdi-download" class="deleted-action" @click="getDeletedHandler" v-if="needShowGetDeletedAction && tokenType === 1">get deleted</q-btn>
           </q-popover>
         </q-btn>
         <div v-if="!items.length">{{isLoading ? 'Fetching data..' : 'Devices not found'}}</div>
@@ -70,6 +71,7 @@
                 </q-item>
               </VirtualList>
             </q-list>
+            <q-btn icon="mdi-download" class="deleted-action" @click="getDeletedHandler" v-if="needShowGetDeletedAction && tokenType === 1">get deleted</q-btn>
           </q-popover>
         </q-item>
         <q-btn v-if="!selectedItem.deleted" flat class="on-left" color="white" @click="modeModel = !modeModel" :icon="modeModel ? 'playlist_play' : 'history'"  :rounded="$q.platform.is.mobile">
@@ -118,6 +120,7 @@
           ref="logs"
           :mode="mode"
           :item="selectedItem"
+          :limit="limit"
           originPattern="gw/devices/:id"
           :isEnabled="!!+size[0]"
           v-if="+size[0]"
@@ -156,7 +159,7 @@
 import logs from '../logs/Index.vue'
 import messages from './messages/Index.vue'
 import MapComponent from './MapComponent'
-import { mapState } from 'vuex'
+import { mapState, mapActions } from 'vuex'
 import VirtualList from 'vue-virtual-scroll-list'
 
 export default {
@@ -175,7 +178,8 @@ export default {
       isInit: false,
       isVisibleMap: false,
       mapMinimizedOptions: {},
-      siblingHeight: null
+      siblingHeight: null,
+      needShowGetDeletedAction: true
     }
   },
   computed: {
@@ -189,7 +193,8 @@ export default {
         let hasntMessages = this.config.messages && state[this.config.messages.vuexModuleName] && !state[this.config.messages.vuexModuleName].messages.length && this.ratio !== 100,
           hasntLogs = this.config.logs && state[this.config.logs.vuexModuleName] && state[this.config.logs.vuexModuleName].messages && !state[this.config.logs.vuexModuleName].messages.length && this.ratio !== 0
         return hasntMessages && hasntLogs
-      }
+      },
+      tokenType (state) { return state.tokenInfo.access ? state.tokenInfo.access.type : -1 }
     }),
     size () {
       return [this.ratio, 100 - this.ratio]
@@ -218,6 +223,7 @@ export default {
     }
   },
   methods: {
+    ...mapActions(['getDeleted']),
     viewDataHandler (content) {
       this.$emit('view-data', content)
       if (this.isVisibleMap && content['position.latitude'] && content['position.longitude']) {
@@ -249,6 +255,10 @@ export default {
         this.$store.commit(`${this.config.logs.vuexModuleName}/clearMessages`)
       })
         .catch(() => {})
+    },
+    async getDeletedHandler () {
+      await this.getDeleted('devices')
+      this.needShowGetDeletedAction = false
     }
   },
   created () {
@@ -330,5 +340,13 @@ export default {
     color: white;
     padding: 0 3px;
     margin-bottom: 3px;
+  }
+  .deleted-action {
+    width: 100%;
+    color: #999;
+    background-color: #eee;
+    font-size: .7rem;
+    padding-top: 0;
+    padding-bottom: 0
   }
 </style>